requirements.txt
py-trello
requests